+++ frontend/src/components/Settings.js
@@ -23,6 +23,7 @@ import { UiContext } from './UiProvider';
 import { NetworkConfigDialog } from './NetworkConfigDialog';
 import { ChangePasswordDialog } from './ChangePasswordDialog';
 import { debug } from '../App';
+// const debug = true;
 
 function Settings() {
   // WebSocket context
@@ -84,6 +85,8 @@ function MatterbridgeSettings({ matterbridgeInfo, systemInfo }) {
   const [selectedMbLoggerLevel, setSelectedMbLoggerLevel] = useState('Info'); 
   const [logOnFileMb, setLogOnFileMb] = useState(false);
   const [frontendTheme, setFrontendTheme] = useState('dark');
+  const [homePagePlugins, setHomePagePlugins] = useState(localStorage.getItem('homePagePlugins')==='false' ? false : true);
+  const [homePageMode, setHomePageMode] = useState(localStorage.getItem('homePageMode')??'logs');
 
   // Ui context
   const { showSnackbarMessage } = useContext(UiContext);
@@ -157,6 +160,22 @@ function MatterbridgeSettings({ matterbridgeInfo, systemInfo }) {
     document.body.setAttribute("frontend-theme", newTheme);
   };
 
+  // Define a function to handle change home page setup
+  const handleChangeHomePagePlugins = (event) => {
+    const newValue = event.target.checked;
+    if(debug) console.log('handleChangeHomePagePlugins called with value:', newValue);
+    setHomePagePlugins(newValue);
+    localStorage.setItem('homePagePlugins', newValue ? 'true' : 'false');
+  };
+
+  // Define a function to handle change home page setup
+  const handleChangeHomePageMode = (event) => {
+    const newValue = event.target.value;
+    if(debug) console.log('handleChangeHomePageMode called with value:', newValue);
+    setHomePageMode(newValue);
+    localStorage.setItem('homePageMode', newValue);
+  };
+
   return (
     <div className="MbfWindowDiv" style={{ flex: '0 0 auto' }}>
       <div className="MbfWindowHeader">
@@ -185,12 +204,23 @@ function MatterbridgeSettings({ matterbridgeInfo, systemInfo }) {
             </Select>
             <FormControlLabel style={{padding: '0px', margin: '0px'}} control={<Checkbox checked={logOnFileMb} onChange={handleLogOnFileMbChange} name="logOnFileMb" />} label="Log on file:" labelPlacement="start"/>
           </div>
-          <div style={{ display: 'flex', alignItems: 'center', gap: '15px' }}>
+          <div style={{ display: 'flex', alignItems: 'center', gap: '15px', marginTop: '5px' }}>
             <FormLabel style={{padding: '0px', margin: '0px'}} id="frontend-theme-label">Frontend theme:</FormLabel>
             <Select style={{ height: '30px' }} labelId="frontend-theme-label" id="frontend-theme" value={frontendTheme} onChange={handleChangeTheme}>
               <MenuItem value='classic'>Classic</MenuItem>
               <MenuItem value='light'>Light</MenuItem>
               <MenuItem value='dark'>Dark</MenuItem>
+            </Select>
+          </div>
+          <div style={{ display: 'flex', alignItems: 'center', gap: '15px', marginTop: '5px' }}>
+            <FormLabel style={{padding: '0px', margin: '0px'}} id="frontend-home-plugin-label">Home page plugins:</FormLabel>
+            <Checkbox checked={homePagePlugins} onChange={handleChangeHomePagePlugins} name="showPlugins" />
+          </div>
+          <div style={{ display: 'flex', alignItems: 'center', gap: '15px', marginTop: '5px' }}>
+            <FormLabel style={{padding: '0px', margin: '0px'}} id="frontend-home-label">Home page bottom panel:</FormLabel>
+            <Select style={{ height: '30px' }} labelId="frontend-home-label" id="frontend-home" value={homePageMode} onChange={handleChangeHomePageMode}>
+              <MenuItem value='logs'>Logs</MenuItem>
+              <MenuItem value='devices'>Devices</MenuItem>
             </Select>
           </div>
           <div style={{ display: 'flex', alignItems: 'center', gap: '15px', marginTop: '20px' }}>